requirements.txt
protobuf==6.31.1
fastapi==0.109.1
uvicorn==0.27.0
grpcio==1.62.0
grpcio-tools==1.62.0  
pika==1.3.2           
python-socketio==5.11.2
requests